Fix webhook integer overflow - change to BigInteger for Strava IDs

+++ src/db/models/webhook_events.py
@@ -8,7 +8,7 @@ received from Strava. Processing happens asynchronously to ensure fast
 webhook response times.
 """
 
-from sqlalchemy import Column, Integer, String, DateTime, JSON, Enum as SQLEnum
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, JSON, Enum as SQLEnum
 from sqlalchemy.sql import func
 from src.db.models.base import Base
 import enum
@@ -44,13 +44,13 @@ class WebhookEvent(Base):
 
     # Strava event data
     object_type = Column(String, nullable=False)  # "activity" or "athlete"
-    object_id = Column(Integer, nullable=False)  # activity_id or athlete_id
+    object_id = Column(BigInteger, nullable=False)  # activity_id or athlete_id
     aspect_type = Column(String, nullable=False)  # "create", "update", "delete"
-    owner_id = Column(Integer, nullable=False)  # athlete_id who owns the object
-    subscription_id = Column(Integer)  # Strava subscription ID
+    owner_id = Column(BigInteger, nullable=False)  # athlete_id who owns the object
+    subscription_id = Column(BigInteger)  # Strava subscription ID
 
     # Event metadata
-    event_time = Column(Integer)  # Unix timestamp from Strava
+    event_time = Column(BigInteger)  # Unix timestamp from Strava
     updates = Column(JSON)  # Additional data (e.g., {"title": true} for updates)
 
     # Processing status
